Module Bap_strings_unscrambler.Ascii

ASCII Characters

Also provides, different subsets of the Ascii character set, e.g., Ascii.Digits, AScii

module Alpha : sig ... end

Letters

module Alphanum : sig ... end

Letters and Numbers

module Digits : Alphabet

Digits

module Printable : Alphabet

All printable ASCII characters

include Alphabet

val length : int

total number of symbols in the alphabet

val index : char -> int

index x maps x to the n'th symbol of an alphabet, if x is a representation of that symbols, returns a number that is outside of [0,len-1] interval if it is not.