Module Ascii.Alphanum

Letters and Numbers

module Caseless : Alphabet

Caseless Letters

include Alphabet

val length : int

total number of symbols in the alphabet

val index : char -> int

index x maps x to the n'th symbol of an alphabet, if x is a representation of that symbols, returns a number that is outside of [0,len-1] interval if it is not.